feat: add Vim healthcheck
As described in `:h health-dev`.

You can run it with `:checkhealth shellbot` (or just `:checkhealth` in
order to run all health checks) and you'll see output like:

    ==============================================================================
    shellbot: require("shellbot.health").check()

    - WARNING SHELLBOT ("/Users/wincent/.config/nvim/pack/bundle/opt/shellbot/lua/target/release/shellbot") is not executable

or:

    ==============================================================================
    shellbot: require("shellbot.health").check()

    - OK SHELLBOT environment variable is set to an executable

Note the location and naming of the healthcheck module
(`shellbot/health.lua`) is dictated by the conventions spelled out in
`:h health-dev`.